TRUNK-3151 - skip delete if temp folder doesn't exist
git-svn-id: http://svn.openmrs.org/openmrs/branches/1.7.x@29032 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ src/api/org/openmrs/util/OpenmrsClassLoader.java
@@ -535,32 +535,34 @@ public class OpenmrsClassLoader extends URLClassLoader {
 	 * @param libCacheFolder 
 	 */
 	public static void deleteOldLibCaches(File libCacheFolder) {
-		FilenameFilter cacheDirFilter = new FilenameFilter() {
-			
-			@Override
-			public boolean accept(File dir, String name) {
-				return name.endsWith(LIBCACHESUFFIX);
-			}
-		};
-		FilenameFilter lockFilter = new FilenameFilter() {
-			
-			@Override
-			public boolean accept(File dir, String name) {
-				return name.equals("lock");
-			}
-		};
-		File tempLocation = libCacheFolder.getParentFile();
-		File[] listFiles = tempLocation.listFiles(cacheDirFilter);
-		for (File cacheDir : listFiles) {
-			//check if it is a directory, but is not the current lib cache
-			if (cacheDir.isDirectory() && !cacheDir.equals(libCacheFolder)) {
-				// check if its not locked by another running openmrs instance
-				if (cacheDir.list(lockFilter).length == 0) {
-					try {
-						OpenmrsUtil.deleteDirectory(cacheDir);
-					}
-					catch (IOException io) {
-						log.warn("Unable to delete: " + cacheDir.getName());
+		if (libCacheFolder.exists()) {
+			FilenameFilter cacheDirFilter = new FilenameFilter() {
+				
+				@Override
+				public boolean accept(File dir, String name) {
+					return name.endsWith(LIBCACHESUFFIX);
+				}
+			};
+			FilenameFilter lockFilter = new FilenameFilter() {
+				
+				@Override
+				public boolean accept(File dir, String name) {
+					return name.equals("lock");
+				}
+			};
+			File tempLocation = libCacheFolder.getParentFile();
+			File[] listFiles = tempLocation.listFiles(cacheDirFilter);
+			for (File cacheDir : listFiles) {
+				//check if it is a directory, but is not the current lib cache
+				if (cacheDir.isDirectory() && !cacheDir.equals(libCacheFolder)) {
+					// check if its not locked by another running openmrs instance
+					if (cacheDir.list(lockFilter).length == 0) {
+						try {
+							OpenmrsUtil.deleteDirectory(cacheDir);
+						}
+						catch (IOException io) {
+							log.warn("Unable to delete: " + cacheDir.getName());
+						}
 					}
 				}
 			}
